Support a relative path in the socket path
Closes ruby/vscode-rdbg#258

We only support a relative path from a workspace directory
